+++ app/lib/check-date-validity.js
@@ -1,0 +1,12 @@
+const isValidDate = (year, month, day) => {
+  const dateObject = new Date(year, month - 1, day)
+  return (
+    dateObject.getFullYear() === year &&
+      dateObject.getMonth() === month - 1 &&
+      dateObject.getDate() === day
+  )
+}
+
+module.exports = {
+  isValidDate
+}
